Add the generated Java files for service conformance tests.
Add simple Makefile to work with it for now.

Cleanup C++ conformance test. I accidentally created people
messages even when not needed.

R=zerny@google.com

Review URL: https://codereview.chromium.org/973433002

+++ tests/service_tests/performance/java/jni/performance_service_wrapper.cc
@@ -59,7 +59,7 @@ JNIEXPORT jint JNICALL Java_fletch_PerformanceService_echo(JNIEnv*, jclass, jint
 
 static void Unwrap_int32_8(void* raw) {
   char* buffer = reinterpret_cast<char*>(raw);
-  int result = *reinterpret_cast<int*>(buffer + 48);
+  int64_t result = *reinterpret_cast<int64_t*>(buffer + 48);
   jobject callback = *reinterpret_cast<jobject*>(buffer + 32);
   JavaVM* vm = *reinterpret_cast<JavaVM**>(buffer + 56);
   JNIEnv* env = attachCurrentThreadAndGetEnv(vm);
